+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added a prelude
 - Documentation fixes and improvements
 - Allow configuring a loading state multiple times (fixes [#60](https://github.com/NiklasEi/bevy_asset_loader/issues/60))
+- Support optional dynamic folder and files assets (fixes [#49](https://github.com/NiklasEi/bevy_asset_loader/issues/49))
 
 ## v0.11.0
 - Support progress tracking through `iyes_progress` ([#6](https://github.com/NiklasEi/bevy_asset_loader/issues/6))